LPS-90950 Move all statement to a new package
